+++ apps/rates-service/deploy.ts
@@ -1,30 +1,14 @@
-import { copyEnvVarsLocal, copyNpmTokenHere } from "../../tools/predeploy";
+import { gCloudDeploy, removeOldAppVersions, SetGCloudConfig } from "../../tools/predeploy";
+import { predeploy } from "./predeploy";
 
-(async () => {
-  // await SetGCloudConfig("GCLOUD_RATES_CONFIG");
-  await copyEnvVarsLocal(
-    "app.secrets.yaml",
-    { RUNTIME_ENV: "gcloud" },
-    [
-      // Required for updating oracle
-      // "InfuraProjectId",
-      // Required for emailing-on-error
-      // "MailjetApiKey",
-      // "MailjetApiSecret",
-      // Required for Finhub
-      // "FinhubApiKey",
-      // Required for market-status
-      // "TradierApiKey"
-    ]
-  );
-  await copyNpmTokenHere(new URL(import.meta.url));
-  // await gCloudDeploy();
-  // // Clean-up after
-  // await removeOldAppVersions();
+await predeploy();
+await SetGCloudConfig("GCLOUD_RATES_CONFIG");
+await gCloudDeploy();
+// Clean-up after
+await removeOldAppVersions();
 
   // Don't forget to deploy our CRON file
   // if (process.env.SETTINGS != 'beta') {
   //   await ShellCmd("gcloud app deploy cron.yaml")
   // }
-})();