+++ apps/tx-processor/src/initialize.ts
@@ -9,7 +9,7 @@ import { SendMail } from "@thecointech/email";
 import { weSellAt, fetchRate } from '@thecointech/fx-rates';
 import { toHuman } from "@thecointech/utilities";
 import { sleep } from '@thecointech/async';
-import { Signer, formatEther, parseUnits } from "ethers";
+import { Signer, formatEther } from "ethers";
 
 export async function initialize() {
   log.debug(' --- Initializing processing --- ');
@@ -42,40 +42,37 @@ export async function initialize() {
 // Verify we have enough gas to run processing
 async function verifyEtherReserves(signer: Signer) {
   const signerBalance = await signer.provider?.getBalance(signer) ?? 0n;
-  const minimumBalance = parseUnits('0.2', "ether");
-  log.debug({ Balance: formatEther(signerBalance) }, "Processing with eth reserves: {Balance}")
-  if (signerBalance < minimumBalance) {
-    log.error(
-      { Balance: formatEther(signerBalance), MinimumBalance: formatEther(minimumBalance), Signer: 'BrokerCAD' },
-      `Signer {Signer} ether balance too low {Balance} < {MinimumBalance}`
-    );
-    await SendMail(`WARNING: Signer balance too low ${formatEther(signerBalance)}`, `Signer balance too low ${formatEther(signerBalance)}\nMinimum balance required: ${minimumBalance}`);
-    await maybeSleep();
-  }
+  const signerAddress = await signer.getAddress();
+  const balanceEth = formatEther(signerBalance);
+  await verifyMinBalance(Number(balanceEth), 0.2, "BrokerCAD", signerAddress, "ether");
 }
 
 // Verify we have enough reserves to run processing
 async function verifyCoinReserves(signer: Signer, contract: TheCoin) {
   const signerAddress = await signer.getAddress();
-  const reservesCoin = await contract.balanceOf(signerAddress);
+  const balanceCoin = await contract.balanceOf(signerAddress);
   const now = new Date();
   const rate = await fetchRate(now);
-  if (!rate) {
-    log.error("tx-processor couldn't fetch current rate");
+  if (!rate?.sell || !rate?.fxRate) {
+    log.fatal("tx-processor couldn't fetch current rate");
     return;
   }
-  const reservesCad = toHuman(
-    Number(reservesCoin) * weSellAt([rate], now),
+  const balanceCad = toHuman(
+    Number(balanceCoin) * weSellAt([rate], now),
     true
   );
-  log.debug({ Balance: reservesCad }, "Processing with $ reserves: {Balance}")
-  const minimumBalance = 10_000;
-  if (reservesCad < minimumBalance) {
+  await verifyMinBalance(balanceCad, 10_000, "BrokerCAD", signerAddress, "$CAD");
+}
+
+async function verifyMinBalance(Balance: number, MinimumBalance: number, Signer: string, Address: string, currency: string) {
+  log.debug({ Balance }, `Processing with ${currency} reserves: {Balance}`)
+  if (Balance < MinimumBalance) {
     log.error(
-      { Balance: reservesCad, MinimumBalance: minimumBalance, Signer: 'BrokerCAD' },
-      `Signer {Signer} $ balance too low {Balance} < {MinimumBalance}`
+      { Balance, MinimumBalance, Signer, Address },
+      `Signer {Signer} ${currency} balance too low {Balance} < {MinimumBalance}`
     );
-    log.error(`Not enough coin reserves: ${formatEther(reservesCad)} CAD`);
+    await SendMail(`WARNING: Signer balance too low ${Balance}`, `${Signer} balance too low ${currency} ${Balance}\nMinimum balance required: ${MinimumBalance}`);
+    log.error(`Not enough ${currency} reserves: ${Balance}`);
     await maybeSleep();
   }
 }