+++ tools/initDemoAccount.ts
@@ -21,7 +21,7 @@ if (process.env.CONFIG_NAME == "devlive") {
 }
 
 /////////////////////////////////////////////
-const monthsToRun = 9999;
+const monthsToRun = 100;
 /////////////////////////////////////////////
 
 // Init/Demo account
@@ -47,7 +47,7 @@ const initBlock = parseInt(process.env.INITIAL_COIN_BLOCK ?? "0", 10);
 const tx = await loadAndMergeHistory(initBlock, tcCore, testAddress);
 // We start from the last deposit transaction
 const deposits = tx.filter(tx => tx.change > 0);
-const lastTxDate = deposits[deposits.length - 1]?.date ?? startDate;
+const lastTxDate = deposits[deposits.length - 1]?.date ?? startDate.minus({hour: 2});
 console.log(`Last tx: ${lastTxDate.toLocaleString(DateTime.DATETIME_SHORT)}`);
 
 const pausedDate = lastTxDate.plus({ hour: 1});
@@ -139,8 +139,6 @@ while (currDate < endDate) {
 
       if (currDate >= pausedDate) {
 
-        console.log(`Sending BillPayment `);
-
         const dueDate = nextPayDate.plus(visaDuePeriod);
         const billPayment = await BuildUberAction(
           mockPayee,
@@ -150,6 +148,10 @@ while (currDate < endDate) {
           CurrencyCode.CAD,
           dueDate
         )
+        const signedAt = DateTime.fromMillis(billPayment.transfer.signedMillis).toLocaleString(DateTime.DATETIME_SHORT);
+        const dueAt = DateTime.fromMillis(billPayment.transfer.transferMillis).toLocaleString(DateTime.DATETIME_SHORT);
+        console.log(`Sending BillPayment: Signed ${signedAt} - Due ${dueAt}`);
+
         await payBillApi.uberBillPayment(billPayment);
       }